Feature/task status history

+++ ToDoList.Application/Commands/TaskListCommand/Delete/DeleteTaskListCommandHandler.cs
@@ -7,7 +7,7 @@ using ToDoList.Domain.Result;
 
 namespace ToDoList.Application.Commands.TaskListCommand.Delete
 {
-    public class DeleteTaskListCommandHandler : ICommandHandler<DeleteTaskListCommand, bool>
+    public sealed class DeleteTaskListCommandHandler : ICommandHandler<DeleteTaskListCommand, bool>
     {
         private readonly IUnitOfWork _unitOfWork;
 
@@ -21,9 +21,36 @@ namespace ToDoList.Application.Commands.TaskListCommand.Delete
             try
             {
                 var response = await _unitOfWork.TaskListRepository.FindByConditions(x => x.Id == request.Id, cancellationToken)
-                    .Result.FirstOrDefaultAsync();
+                    .Result.Include(x => x.SingleTask).FirstOrDefaultAsync();
 
-                 await _unitOfWork.TaskListRepository.DeleteAsync(response);
+                var newList = await _unitOfWork.TaskListRepository.FindByConditions(x => x.Id == request.NewListId, cancellationToken)
+                    .Result.Include(x => x.SingleTask).FirstOrDefaultAsync();
+
+                if (response is null)
+                {
+                    return new BaseResult<bool>
+                    {
+                        ErrorCode = (int)ErrorCodes.TaskListNotFound,
+                        ErrorMessage = "Список не найден",
+                    };
+                }
+
+                if (!response.SingleTask.Any())
+                {
+                    await _unitOfWork.TaskListRepository.DeleteAsync(response);
+
+                }
+                else
+                {
+                    foreach (var singleTask in response.SingleTask.ToList())
+                    {
+                        newList.SingleTask.Add(singleTask);
+                    }
+                    await _unitOfWork.TaskListRepository.UpdateAsync(newList);
+                    await _unitOfWork.SaveChangesAsync(cancellationToken);
+                    await _unitOfWork.TaskListRepository.DeleteAsync(response);
+                }
+
                 await _unitOfWork.SaveChangesAsync(cancellationToken);
 
                 return new BaseResult<bool>
